requirements.txt
autogluon.tabular>=1.0
groq>=0.5.0
hdbscan==0.8.33
keyring==25.1.0
pandas>=1.5.3
PyYAML==6.0.1
requests_oauthlib==2.0.0
scikit-learn>=1.3.0
sentence_transformers==2.6.1
streamlit>=1.30
plotly>=5.18
